pyproject.toml
[project]
name = "triumvirlabs"
version = "0.1.0"
description = "ElevenLabs Worldwide Hackathon"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "streamlit (>=1.52.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "elevenlabs (>=2.26.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
